Split base bridge.py
